pyproject.toml
[project]
name = "modi"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "haruiz", email = "henryruiz22@gmail.com" } 
]
requires-python = ">=3.10"
dependencies = [
    "anthropic>=0.49.0",
    "google-genai>=1.7.0",
    "groq>=0.11.0",
    "nltk>=3.9.1",
    "ollama>=0.3.3",
    "openai>=1.68.2",
    "pillow>=11.0.0",
    "pydantic>=2.9.2",
    "python-dotenv>=1.0.1",
    "tenacity>=9.0.0",
]

[project.scripts]
modi = "modi:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
